星期天

+++ src/components/admin/EightModuleDetail.vue
@@ -13,9 +13,9 @@
                 <Col span="22">
                     <div class="titletopic">                      
                         <span> <Icon type="ios-book" color="#8DC143" />{{member.memberTitle}}</span> 
-                        <Button class="button1" type="primary" shape="circle" icon="md-add">新增八股</Button>
+                        <Button @click="addNewTopic" class="button1" type="success"  shape="circle" icon="md-add">新增八股</Button>
                     </div>
-                    <Table border :columns="columns12" :data="data6">
+                    <Table height="520" border :columns="columns12" :data="data6" class="table">
                         <template slot-scope="{ row }" slot="topicText">
                             <strong>{{ row.topicText }}</strong>
                         </template>
@@ -25,10 +25,10 @@
                         </template>
 
                         <template slot-scope="{row,index }" slot="action">
-                            <Button type="success" size="small" style="margin-right: 5px" ghost>编辑题目</Button>
+                            <Button type="success" size="small" style="margin-right: 5px" ghost @click="EditEightTitle(row,index)">编辑题目</Button>
                             <Button type="success" size="small" style="margin-right: 5px" @click="showeditor(row,index)">查看</Button>
                             <Button type="primary" size="small" style="margin-right: 5px" @click="EditEight(row,index)">编辑</Button>
-                            <Button type="error" size="small" @click="DeleteEight(row,index)">删除</Button>
+                            <Button type="error" size="small" @click="DeleteEightTopic(row,index)">删除</Button>
                         </template>
                    </Table>
                     <div class="mypage">
@@ -40,15 +40,13 @@
              </Row>
           </div>
 
-          <Modal class="modal" :styles="{top: '1px'}" footer-hide title="士大夫士大夫胜多负少"   width="50vw" v-model="modal10">
+          <Modal class="modal" :styles="{top: '1px'}" footer-hide :title="showRow.topicText"   width="50vw" v-model="modal10">
                <div class="centerdiv">
                   <v-md-editor :value="markdown" mode="preview" class="editor" ></v-md-editor>
                </div>
           </Modal>
 
-          <Modal v-model="modal11" fullscreen :title="nowTopic.topicText"  
-          :loading="loading"
-           @on-ok="asyncOK">
+          <Modal v-model="modal11" fullscreen :title="nowTopic.topicText">
             <div class="modalediteight">
                 <v-md-editor v-model="markdownText" class="editor" :left-toolbar="leftToolbar" height="100%" >
                 </v-md-editor>
@@ -66,6 +64,48 @@
            
             </div>
          </Modal>
+
+            <Modal
+                v-model="modal6"
+                title="题目编辑">
+                <Form :model="formItem" :label-width="80">
+                        <FormItem label="题目">
+                            <Input v-model="formItem.topicText" maxlength="16" show-word-limit placeholder="请输入题目..." ></Input>
+                        </FormItem>
+                </Form>   
+                <div slot="footer">
+                <Row >
+                  <Col span="8"></Col>
+                  <Col span="8" class="col6">
+                    <div class="col6-div">
+                        <Button @click="ExitEditEditEightTitle">取消</Button>   <Button @click="ConfirmSaveEditEightTitle" :loading="loading1" type="primary" >保存</Button>
+                    </div>
+                  </Col>
+                  <Col span="8"></Col>
+                </Row>
+           
+            </div>
+            </Modal>
+
+
+            <Modal v-model="addTopic" title="新增八股文题目" >
+                <Form :model="newTopic" :label-width="80">
+                        <FormItem label="题目">
+                            <Input v-model="newTopic.topicText" maxlength="16" show-word-limit placeholder="请输入题目..." ></Input>
+                        </FormItem>
+                </Form>  
+                <div slot="footer">
+                 <Row >
+                    <Col span="8"></Col>
+                    <Col span="8" class="col6">
+                        <div class="col6-div">
+                            <Button @click="ExitNewTopic">取消</Button>   <Button @click="ConfirmSaveNewTopic" :loading="loading2" type="primary" >保存</Button>
+                        </div>
+                    </Col>
+                    <Col span="8"></Col>
+                </Row>
+                </div>
+           </Modal>
     </div>
 </template>
 
@@ -79,7 +119,7 @@ export default {
             leftToolbar: "h bold italic strikethrough quote | ul ol table hr | link image code | undo redo clear | save |  emoji", // 左侧工具栏
             rightToolbar: "preview toc sync-scroll fullscreen", // 右侧工具栏
             PageSize:1,
-            spinShow:true,
+            spinShow:false,
             value: true,
             member:{
             memberTitle: "",
@@ -105,7 +145,7 @@ export default {
                         align: 'center'
                     },
                     {
-                        title: 'Action',
+                        title: '操作',
                         slot: 'action',
                         width: 300,
                         align: 'center'
@@ -113,10 +153,23 @@ export default {
                 ],
             data6: [],
             modal10:false,
-            markdown:"sdfgsfgsfdgsfdghjfjfghjj",
+            markdown:"",
             modal11:false,
             markdownText:"",
             nowTopic:"",
+            showRow:"",
+            modal6:false,
+            loading1:false,
+            formItem:{
+                topicText:"",
+            },
+            nowTopicIndex:-1,
+            addTopic:false,
+            newTopic:{
+                topicText:"",
+                memberId:"",
+            },
+            loading2:false,
             
            };
     },
@@ -132,14 +185,9 @@ export default {
 
     methods: {
         backpage(){
-            this.$router.back();
+            this.$router.push("/interviewadmin/eightmanagement");
         },
-        show (index) {
-           this.$Modal.info({
-             title: 'User Info',
-             content: "lkhkjh"
-             })
-        },
+       
        isOpenSwitch(row, index){
         this.spinShow = true;      
         this.$axios({
@@ -165,26 +213,41 @@ export default {
        remove (index) {
      
         },
-       DeleteEight(row,index){
-        this.$Message.success('保存sdf成功');    
-        this.$axios({
-                method: "DELETE",
-                url: "/interview/deletetopic",
-                params: {
-                    topicId: row.topicId,
-                },
-                }).then((res) => {
-                    if(res.data.code == 200){
-                                  
-                    }else{
-                     
+       DeleteEightTopic(row,index){
+        this.$Modal.confirm({
+                    title: '删除确认',
+                    content: '<p>确定删除吗？</p>',
+                    onOk: () => {
+                        this.spinShow = true;                      
+                        this.$axios({
+                        method: "DELETE",
+                        url: "/interview/deletetopic",
+                        params: {
+                            topicId: row.topicId,
+                        },
+                        }).then((res) => {
+                            if(res.data.code == 200){
+                                this.data6.splice(index,1);
+                                this.$Message.success("删除成功");
+                            }else{
+                                this.$Message.error("删除失败");
+                            }
+                            this.spinShow = false;
+                        }).catch((res) => {
+                            this.spinShow = false;
+                            this.$Message.error("删除失败");
+                            
+                        });   
+                    },
+                    onCancel: () => {
+                        this.spinShow = false;       
                     }
-                    
-                }).catch((res) => {
-                    this.$Message.error("查询失败");
-                });   
+                });
+       
+       
        },
         showeditor(row,index){
+           this.showRow = row;
          this.$axios({
                 method: "GET",
                 url: "/interview/getTopicanswer",
@@ -196,8 +259,9 @@ export default {
                         this.markdown = res.data.data.articleText;
                         this.modal10 = true;
                     }else{
-                    this.$Message.error("查询失败");
-                }
+                    this.markdown = "还没有内容";
+                    this.modal10 = true;
+                    }
                     
                 }).catch((res) => {
                     this.$Message.error("查询失败");
@@ -205,6 +269,7 @@ export default {
          },
          EditEight(row,index){
            this.nowTopic = row;
+           this.modal11 = true;
            this.$axios({
                 method: "GET",
                 url: "/interview/getTopicanswer",
@@ -212,12 +277,11 @@ export default {
                     topicId: row.topicId,
                 },
                 }).then((res) => {
-                    if(res.data.code == 200){
-                        this.modal11 = true;
+                    if(res.data.code == 200){                     
                         this.markdownText = res.data.data.articleText;                      
                     }else{
-                    this.$Message.error("查询失败");
-                }
+                     this.markdownText = "还没有内容";
+                   }
                     
                 }).catch((res) => {
                     this.$Message.error("查询失败");
@@ -262,7 +326,7 @@ export default {
               url: "/interview/getpagesize",
               params: {
                 memberId:this.member.memberId
-              },       
+              },
             }).then((res) => {             
                 if(res.data.code == 200){ 
                    this.PageSize = res.data.data;
@@ -287,16 +351,81 @@ export default {
                 if(res.data.code == 200){ 
                     this.$Message.success('保存成功');                 
                 }else{                  
-                    this.$Message.error('保存');
+                    this.$Message.error('保存失败');
                 }  
                 this.loading = false;                                       
               }).catch((res) => {     
                 this.loading = false;  
-                this.$Message.error('保存');
+                this.$Message.error('保存失败');
               });
         },
         ExitEditing(){
            this.modal11 = false;
+        },
+        EditEightTitle(row,index){
+            this.formItem =  JSON.parse(JSON.stringify(row));
+            this.modal6 = true;
+            this.nowTopicIndex = index;
+        },
+       
+        ConfirmSaveEditEightTitle(){
+            this.loading1 = true;
+            this.$axios({
+              method: "PUT",
+              url: "/interview/saveediteeighttitle",
+              data: {
+                topicId:this.formItem.topicId,
+                topicText:this.formItem.topicText,
+              },
+            }).then((res) => {              
+                if(res.data.code == 200){ 
+                    this.$Message.success('更改成功');  
+                    this.data6[this.nowTopicIndex].topicText = this.formItem.topicText;
+                }else{                  
+                    this.$Message.error('更改失败');
+                }  
+                this.loading1 = false;    
+                this.modal6 = false;                                   
+              }).catch((res) => {     
+                this.loading1 = false;
+                this.$Message.error('更改失败');
+                this.modal6 = false; 
+              });
+        },
+        ExitEditEditEightTitle(){
+            this.modal6 = false; 
+        },
+        addNewTopic(){
+            this.addTopic = true;
+            this.newTopic.memberId = this.member.memberId;
+            this.newTopic.topicText = "";
+        },
+        ConfirmSaveNewTopic(){
+            this.loading2 = true;
+            this.$axios({
+              method: "POST",
+              url: "/interview/savenewtopic",
+              data: this.newTopic
+            }).then((res) => {              
+                if(res.data.code == 200){ 
+                    this.$Message.success('增加成功');  
+                    this.data6.unshift(res.data.data);
+                }else{                  
+                    this.$Message.error('增加失败');
+                }  
+                this.loading2 = false;    
+                this.addTopic = false;
+                                                
+              }).catch((res) => {     
+                this.loading2 = false;
+                this.addTopic = false;
+                this.$Message.error('增加失败');
+               
+              });
+
+        },
+        ExitNewTopic(){
+            this.addTopic = false;
         }
     },
 };
